+++ speech/deepgram/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @mastra/speech-deepgram
 
+## 0.1.7-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [dab255b]
+  - @mastra/core@0.5.0-alpha.1
+
+## 0.1.7-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [59df7b6]
+- Updated dependencies [29f3a82]
+- Updated dependencies [59df7b6]
+- Updated dependencies [c139344]
+  - @mastra/core@0.5.0-alpha.0
+
 ## 0.1.6
 
 ### Patch Changes